data sdk: use ip address instead of partition id to rule out unavailable data partitions
In sequential write protocol, a write is considered successful only if
all of the partition members reply success. So using ip address instead
of partition id to rule out unavailable data partitions would be more
appropriate, i.e. any member in a data partition should not be in the
exclude list.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>